// src/types/bookItem.ts
// export default interface bookItem {
//     workbook_id: string;
//     title: string;
//     img: string;
//     publisher: string;
//     level: number;
//     like: number;
// }

export default interface bookItem {
    workbookId: string;
    title: string;
    profileImg: string;
    publisher: string;
    level: number;
    like: number;
}